pyproject.toml
[tool.poetry]
name = "le-wagon-mlops-optimization-taxifare-api"
version = "0.1.0"
description = "api package to the Taxifare Opts project"
authors = ["Arielle Villa Massone <arielle.villamassone@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.115.4"
gunicorn = "^23.0.0"
uvicorn = "^0.32.0"
pandas = "^1.3"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.18.2"
scikit-learn = "^1.0"
pygeohash = "^1.2.0"
mlflow = "^2.17.2"
pyspark = "^3.5.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
